Ton Hospel made a couple interesting observations about POE's event loop.  First, the code could be tightened up.  Perl assignments are relatively expensive, and restructuring the code to avoid them means a few percent increase in performance.  His example appears as a comment in POE::Queue::Array, marked with his name.
His second observation is that the Tk loop is about twice as slow as
it should be.  Our workaround for high-priority Tk timers is rather
stupid.  Ton works around it by maintaining separate "immediate" and
timer queues, and not using Tk to drive the immediate events.  I've
commented on his idea in POE::Loop::TkCommon.


git-svn-id: https://poe.svn.sourceforge.net/svnroot/poe/trunk@1623 742d25e2-4c0f-0410-ba8b-f6b7b6679366