Change signal watchers so they keep sessions alive.
WARNING: This is a major semantics change in POE. It has the potential to make code 'hang' in places where it formerly did not.

This change is necessary so sessions expressing an interest in SIG CH?LD do not die prematurely. (There is a planned mandatory warning for reaped children that were not being watched.)
This change fixes RT 15215.


git-svn-id: https://poe.svn.sourceforge.net/svnroot/poe/trunk@1852 742d25e2-4c0f-0410-ba8b-f6b7b6679366