Confess when something fails, rather than merely die.  Many generic functions died with immediate failure causes, but they didn't really tell us what was happening at the time.
For example, keep files weren't being found, and the basic read
function was dying.  All we could tell was that a keep file read was
failing but not what part of the wiki triggered the read.  Once we had
the call stack, we knew where to put a -f test and use default data.


git-svn-id: https://poe.svn.sourceforge.net/svnroot/poe/trunk@2573 742d25e2-4c0f-0410-ba8b-f6b7b6679366