!!! This change breaks backward compatibility on a relatively unused !!! feature.  You are affected if you use ARG0 or ARG1 in a !!! POE::Component::Server::TCP ClientConnected callback.
ClientArgs promised more than it could deliver, and people finally
noticed.  This change backs off supplying the socket in $_[ARG0], and
it expands ClientArgs' arrayref into @_[ARG0..$#_].  Thanks to Michael
Fowler for rt.cpan.org #47855 (which this resolves), and POE's mailing
list for advice on which way this change should go.


git-svn-id: https://poe.svn.sourceforge.net/svnroot/poe/trunk@2613 742d25e2-4c0f-0410-ba8b-f6b7b6679366