Fix an infinite file rollover when open() succeeds but stat() fails.
Jay Hannah discovered a bug when he passed in a filename with trailing
whitespace.  Two-parameter open() treats trailing whitespace as
insignificant, so it passed.  On the other hand, trailing whitespace
is significant in stat(), so it always failed, triggering an
unrelenting stream of file rollover events.  Switched to
three-parameter open(), which considers trailing whitespace as
significant.  Now both open() and stat() agree the file doesn't exist.

Thanks to Jay and the University of Nebraska at Omaha for securing me
some debugging time on their machines.


git-svn-id: https://poe.svn.sourceforge.net/svnroot/poe/trunk@2816 742d25e2-4c0f-0410-ba8b-f6b7b6679366